refactor(expand): move support types under expand.ts
