+++ src/operator/expand.ts
@@ -1,6 +1,12 @@
 import {Observable} from '../Observable';
 import {Scheduler} from '../Scheduler';
-import {ExpandOperator} from './expand-support';
+import {Operator} from '../Operator';
+import {Subscriber} from '../Subscriber';
+import {tryCatch} from '../util/tryCatch';
+import {errorObject} from '../util/errorObject';
+import {Subscription} from '../Subscription';
+import {OuterSubscriber} from '../OuterSubscriber';
+import {subscribeToResult} from '../util/subscribeToResult';
 
 /**
  * Returns an Observable where for each item in the source Observable, the supplied function is applied to each item,
@@ -16,4 +22,93 @@ export function expand<T, R>(project: (value: T, index: number) => Observable<R>
   concurrent = (concurrent || 0) < 1 ? Number.POSITIVE_INFINITY : concurrent;
 
   return this.lift(new ExpandOperator(project, concurrent, scheduler));
+}
+
+export class ExpandOperator<T, R> implements Operator<T, R> {
+  constructor(private project: (value: T, index: number) => Observable<R>,
+              private concurrent: number,
+              private scheduler: Scheduler) {
+  }
+
+  call(subscriber: Subscriber<R>): Subscriber<T> {
+    return new ExpandSubscriber(subscriber, this.project, this.concurrent, this.scheduler);
+  }
+}
+
+export class ExpandSubscriber<T, R> extends OuterSubscriber<T, R> {
+  private index: number = 0;
+  private active: number = 0;
+  private hasCompleted: boolean = false;
+  private buffer: any[];
+
+  constructor(destination: Subscriber<R>,
+              private project: (value: T, index: number) => Observable<R>,
+              private concurrent: number,
+              private scheduler: Scheduler) {
+    super(destination);
+    if (concurrent < Number.POSITIVE_INFINITY) {
+      this.buffer = [];
+    }
+  }
+
+  private static dispatch({subscriber, result, value, index}): void {
+    subscriber.subscribeToProjection(result, value, index);
+  }
+
+  protected _next(value: any): void {
+    const destination = this.destination;
+
+    if (destination.isUnsubscribed) {
+      this._complete();
+      return;
+    }
+
+    const index = this.index++;
+    if (this.active < this.concurrent) {
+      destination.next(value);
+      let result = tryCatch(this.project)(value, index);
+      if (result === errorObject) {
+        destination.error(errorObject.e);
+      } else if (!this.scheduler) {
+        this.subscribeToProjection(result, value, index);
+      } else {
+        const state = { subscriber: this, result, value, index };
+        this.add(this.scheduler.schedule(ExpandSubscriber.dispatch, 0, state));
+      }
+    } else {
+      this.buffer.push(value);
+    }
+  }
+
+  private subscribeToProjection(result: any, value: T, index: number): void {
+    if (result._isScalar) {
+      this._next(result.value);
+    } else {
+      this.active++;
+      this.add(subscribeToResult<T, R>(this, result, value, index));
+    }
+  }
+
+  protected _complete(): void {
+    this.hasCompleted = true;
+    if (this.hasCompleted && this.active === 0) {
+      this.destination.complete();
+    }
+  }
+
+  notifyNext(outerValue: T, innerValue: R, outerIndex: number, innerIndex: number): void {
+    this._next(innerValue);
+  }
+
+  notifyComplete(innerSub: Subscription): void {
+    const buffer = this.buffer;
+    this.remove(innerSub);
+    this.active--;
+    if (buffer && buffer.length > 0) {
+      this._next(buffer.shift());
+    }
+    if (this.hasCompleted && this.active === 0) {
+      this.destination.complete();
+    }
+  }
 }
